Aux

Documentation

tupToList :: (a, a) -> [a]

Converts a tuple to a list

lastTup :: [a] -> Maybe ([a], a)

Illustrative example:

>>> lastTup [1,2,3,4]
([1,2,3], 4)

Used to parse input vars, ex:

lastTup ["unsigned", "int", "x"] == Just (["unsigned", "int"], "x")

apFst :: (a1 -> a2) -> (a1, a) -> (a2, a)

Applies a function to the first element of a tuple.

apToLast :: (a -> a) -> [a] -> [a]

Apply the input function to the last element of the list

>>> apToLast (const 0) [1,2,3,4]
[1,2,3,0]