CGen.Typed

Documentation

addUnsigned :: Text -> Text

Adds U to the end of a Text object

addLong :: Text -> Text

Adds L to the end of a Text object

showt :: Show a => a -> Text

class CShow a where

This class allows one to show objects from Foreign.C.Types in a way that won't throw warnings if used in C code. For example, 'unsigned long long' literals need the ULL annotation.

Methods

showForC :: a -> Text