Data.Either.Aux

Documentation

fromLeft :: Either a b -> Maybe a

This provides safe unpacking of a Left object.

fromRight :: Either a b -> Maybe b

This provides safe unpacking of a Right object.

mapLeft :: (t -> a) -> [Either t b] -> [Either a b]

This maps a function over a list, only applying it to Left objects

mapRight :: (t -> b) -> [Either a t] -> [Either a b]

This maps a function over a list, only applying it to Right objects

unsafeFromLeft :: Either a b -> a

This will throw an error if its input is Right