Data.Text.Aux

Documentation

showInt :: Int -> Text

showInt should be equivalent to `pack . show` for the Int type, except a bit faster since it uses Text library functions

textAp :: String -> Text -> Text

Example:

>>> textAp "Hello" "World"
"Hello (World)"

wrapText :: Char -> Text -> Text

Example:

>>> wrapText '|' "text"
"|text|"

addArrows :: [Text] -> Text

Example:

>>> addArrows . map T.pack ["Bool", "String", "Int"]
"Bool -> String -> Int"

unwords2 :: Text -> Text -> Text

Example:

>>> unwords2 (T.pack "Hello") (T.pack "World")
"Hello World"

appendAfter :: Text -> Text -> Text -> Text

Example:

>>> appendAfter ", " "Hello" "World"
"Hello, World"

parens :: Text -> Text

Add parentheses around a Text object

wordLines :: Text -> [[Text]]

wordLines converts a Text input to a list of lists, where `wordLines text !! i !! j` is the jth word of the ith line of text.

unWordLines :: [[Text]] -> Text

See wordLines

packUnlines :: [String] -> Text

Shortcut for

unlines . map pack

packUnwords :: [String] -> Text

Shortcut for

unwords . map pack

prettyShowt :: Pretty p => p -> Text

Convert a pretty-printable object to Text