Data.Wrapped

Documentation

data Wrapped a b

This is a wrapper type made specifically for Compilable instances. a should be a functional type and b the "last" return type of that function. In all instances, a is considered a kind of metadata (as it should only be accessed by compiling functions).

Constructors

Wrap a b

defWrap :: Default a => b -> Wrapped a b

unwrapF :: Wrapped a b -> a

Unwrap the function

unwrapR :: Wrapped a b -> b

Unwrap the return value

unwrap :: Wrapped a b -> b

Synonym for unwrapR

wrappedAp :: Wrapped (t -> a) b -> t -> Wrapped a b

wrappedAp or `wrapped function apply`