HSGen.Recompiler.Class

Documentation

class Resolvable a r | a -> r where

This class is of functions with type a and final return type r. Additionally, r should be of the form: Wrapped a b

Methods

resolve :: a -> r

resolve resolves a "Resolvable a r" to r

class Compilable a r b s | a -> r, b -> s, a s -> b, r b -> a where

Methods

($$$) :: (r -> s) -> a -> b infixr 2

`($$$) f wrappedFunction` essentially applies f directly to the Wrapped part of wrappedFunction.

($$) :: Compilable a1 (Wrapped (t -> a) b) t1 (Wrapped a b) => (t -> a1) -> t -> t1 infixl 1

Given a simple Compilable function:

   plus :: Int -> Int -> Wrapped (Int -> Int -> Int) Int

You can use `($$)` to apply plus to both the outer (Int -> Int -> ..) and innner (Wrapped (Int -> ..)) functions of plus. For example:

>>> plus $$ 1 $$ 2
Wrap 3 3

class FixResolvable a r | a -> r where

See Resolvable

Methods

fixResolve :: a -> r

class FixCompilable a r b s | a -> r, b -> s, a s -> b, r b -> a where

See Compilable

Methods

(###) :: (r -> s) -> a -> b infixr 2

(##) :: FixCompilable a (Wrapped (r -> a1) b) (r -> r1) (Wrapped a1 b) => a -> r -> r1

See `($$)`

apInOut :: Compilable a b b c => (b -> c) -> a -> c

Apply a function to a Compilable from the outside (just regular application) and also inside (`($$$)`)

fixApInOut :: FixCompilable a b b c => (b -> c) -> a -> c

See apInOut

flipC :: Compilable a1 (a -> b -> c) c (b -> a -> c) => (a -> b -> a1) -> b -> a -> c

Flip a compilable function

fixFlipC :: FixCompilable a (a1 -> b -> c) (a1 -> b -> c) (b -> a1 -> c) => a -> b -> a1 -> c

See flipC

replaceFW :: a -> Wrapped a b -> Wrapped a b

compile :: (Compilable a1 (Wrapped a b) r (Wrapped a b), Resolvable a1 (Wrapped a b1)) => a1 -> r

Compile a compilable function, essentially by resolving to the wrapped function and return value, unwrapping the non-compilable function (i.e. getting the compiled version), and replacing the original non-compilable function.

fixCompile :: (Compilable a1 (Wrapped a b) r (Wrapped a b), Resolvable a1 (Wrapped a b1)) => a1 -> r

See compile