HSGen.Recompiler.Instances

Description

Deprecated: All derivation should be covered by Class.hs

Documentation

oneFix :: Int -> Text

Example: `oneFix 13 = Text.pack "Fix a32"`

fixes :: Int -> Text

`fixes 3 == pack "Fix a3 -> Fix a2 -> Fix a1"`

makeFwr :: Int -> Text

`makeFwr 2 == pack "Fix a2 -> Fix a1 -> Wrapped f r"`

makeFws :: Int -> Text

`makeFws 2 == pack "Fix a2 -> Fix a1 -> Wrapped g s"`

wr :: Text

"Wrapped f r"

ws :: Text

"Wrapped g s"

makeXs :: Int -> Text

makeUnfixeds :: Int -> Text

makeLambdaXs :: Int -> Text

`makeLambdaXs 3 == pack "\x1 x2 x3 ->"`

genInstance :: Int -> Text

Generate the text for an instance of the class Compilable with n arguments of the form `Fix ai`

classStr :: [String]

Raw text used by classText

classText :: Text

Text form of Compilable